refactor(types): вынести ConfigurationGenericExpander для S104/S1448
